Expand entrypoint target kind options but make them optional (Deprecated).
This parameter is not used in composer.
Composer has to infer the type anyway to be able to type check it,
which just increases the amount of code. Currently it always infers it.
However for completeness, expanded the options and kept the feature
because many blueprints (in tests or example apps) use it.