add CLI DB migrations
Prisma requires migrations to be in the same foldera as prisma schema file. Since migrations have to be source controlled we need to move both schema and migrations to client's project root. This commit introduces "gaudi" folder which should contain gaudi created files that need to be source controlled by clients.